uptime/requirements.txt
functions-framework>=3.0.0
google-cloud-firestore>=2.11.0
google-auth>=2.17.0
bittensor==9.12.2

